LPS-56848 Move to internal package
